% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fast.TLE.BST.R
\name{fast.TLE.BST}
\alias{fast.TLE.BST}
\title{Wrapper to fastTLE assuming bivariate skew-t data.}
\usage{
fast.TLE.BST(data, k, initial = c(0, 0, 1, 0, 1, 0, 0, 1000), ...)
}
\arguments{
\item{data}{The data used in the analysis.}

\item{k}{The proportion of observations used.  If k=1, the (non-robust) MLE
will be computed.}

\item{initial}{A vector of length eight containing an initial guess for
xi1, xi2, omega11, omega12, omega22, alpha1, alpha2, and nu.  Defaults to a
zero mean, identity omega, no skew, and large nu distribution (i.e.
approximately N(0, I)).  Note: do not set the nu parameter to Inf, as this
causes issues with the optimization.}
}
\value{
Same as fastTLE.
}
\description{
Wrapper to fastTLE assuming bivariate skew-t data.
}

