% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fast.TLE.ST.R
\name{fast.TLE.ST}
\alias{fast.TLE.ST}
\title{Wrapper to fastTLE assuming univariate skew-t data.}
\usage{
fast.TLE.ST(data, k, initial = c(0, 1, 0, 1000), ...)
}
\arguments{
\item{data}{The data used in the analysis.}

\item{k}{The proportion of observations used.  If k=1, the (non-robust) MLE
will be computed.}

\item{initial}{A vector of length four containing an initial guess for xi,
omega, alpha, and nu.}
}
\value{
Same as fastTLE.
}
\description{
Wrapper to fastTLE assuming univariate skew-t data.
}

