% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/marginal.R
\name{marginal}
\alias{marginal}
\title{Convert multivariate skew-t parameters into bivariate}
\usage{
marginal(xi, omega, alpha, nu, r = 1, alphaAdj = F)
}
\arguments{
\item{xi}{Vector of parameters for the skew-t distribution.}

\item{omega}{matrix of parameters for the skew-t distribution.}

\item{alpha}{Vector of parameters for the skew-t distribution.}

\item{nu}{Numeric value of parameter for the skew-t distribution.}

\item{r}{This function is currently very specific to the 16x16 case.  r
specifies which of the 8 distributions you wish to see, and then extracts
the (r, r+8) parameters from that distribution.}

\item{alphaAdj}{If TRUE, computes the correct alpha for the marginal (per
Azzalini et.al's "Distributions generated by perturbations of symmetry",
page 18 and Capitano et.al's "Graphical models for skew-normal variates",
page 15).  If false, just subsets alpha.}
}
\value{
A list of the 4 parameters of the bivariate skew-t.
}
\description{
This function converts parameters of a multivariate skew-t distribution
into parameters for a bivariate skew-t using formulas from
wind_radiosonde_QC_01.pdf (from Mandy).
}

