% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mst.pdev.grad.robust.R
\name{mst.pdev.grad.robust}
\alias{mst.pdev.grad.robust}
\title{Gradient of robust, penalized deviance for multivariate skew-t}
\usage{
mst.pdev.grad.robust(param, x, y, k = 2, ...)
}
\arguments{
\item{param}{Optimization parameters derived from xi, Omega, alpha, nu. For
conversion, see the functions optpar2dplist and dplist2optpar.}

\item{x}{A matrix of the independent variables for the fit.  Typically just
a matrix of ones.}

\item{y}{A matrix of dependent variables.}

\item{k}{Parameter controlling the robustness of the fit.  The largest
possible value for the negative log-likelihood is 2*k, and the negative
log-likelihood is adjusted down whenever it is larger than k.}
}
\value{
The gradient of the robust deviance with respect to dp.
}
\description{
This function computes the gradient of the robust-ified, penalized deviance
for the multivariate skew-t (with respect to the optimization parameters).
}

