% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/n.dev.R
\name{n.dev}
\alias{n.dev}
\title{Normal deviance}
\usage{
n.dev(dp, y)
}
\arguments{
\item{dp}{The "density parameters".  More specifically, just a vector of
length two containing the mean and standard deviation parameters.}

\item{y}{A vector of observed values.}
}
\value{
The sum of the negative log-likelihoods of a normal with parameters
dp evaluated at the values y.
}
\description{
Normal deviance
}

