% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/psi.grad.R
\name{psi.grad}
\alias{psi.grad}
\title{Robust adjustment of the derivative of the negative log-likelihood}
\usage{
psi.grad(originalNLL, k)
}
\arguments{
\item{originalNLL}{The original negative log-likelihood}

\item{k}{A parameter controlling the amount of robustification.  Negative
log-likelihoods larger than k are reduced, and the adjusted value is always
less than 2*k.}
}
\value{
Returns
}
\description{
This function implements an adjustment to the derivative of the density
function to make estimation of parameters more robust.  If a negative
log-likelihood is supplied, then psi returns an adjusted negative
log-likelihood. Essentially, if the negative log-likelihood value is too
large, it is bounded via this function.
}
\details{
Usually, we're interested in df/dp, where f is the negative log-likelihood
and p a parameter.  df/dp should be updated with df/dp*psi(f, p) to be
"robustified".
}

