% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Psi.R
\name{Psi}
\alias{Psi}
\title{Robust adjustment of the negative log-likelihood}
\usage{
Psi(originalNLL, k)
}
\arguments{
\item{originalNLL}{The original negative log-likelihood}

\item{k}{The parameter controlling the amount of robustification.  Negative
log-likelihoods larger than k are reduced, and the adjusted value is always
less than 2*k.}
}
\value{
The adjusted negative log-likelihood.
}
\description{
Function implementing the robustness adjustment to the negative
log-likelihood.
}
\details{
Essentially, if the negative log-likelihood value is too large, it is
bounded via Psi().  The psi() function bounds the derivative.
}

