% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/robustST_changing_k.R
\name{robustSTChangingK}
\alias{robustSTChangingK}
\title{Fit a robust Skew-t, iteratively update k}
\usage{
robustSTChangingK(y, x = matrix(1, nrow = NROW(y)), robust = T,
  method = c("nlminb", "constrOptim"), w = rep(1, nrow(x)), pValue = 0.01,
  start = NULL)
}
\arguments{
\item{y}{A vector or matrix of observations to fit the skew-t to.}

\item{x}{A matrix of ones, or matrix of independent variables for skew-t
regression (use caution, as this feature has not been tested!)}

\item{robust}{Should the robust estimator be used?}

\item{w}{A vector of case weights, defaults to a vector of ones.}

\item{pValue}{A parameter controlling the "robustness" of the fit.  Given
current parameter estimates, a (1-pValue)% confidence region can be
constructed, and observations in this region will not be adjusted during
the optimization.  However, values outside this region will have their
likelihood adjusted down, and hence will have less influence on the
M-estimator.  As pValue->1 the estimator approaches the MLE.}

\item{start}{The starting values for the optimization.  If NULL, reasonable
values are automatically chosen.}

\item{method:}{constrOptim uses a constrained algorithm, forcing nu and
omega>0.  However, the implementation for multivariate skew-t fitting
enforces this by default, so nlminb and constrOptim should be very similar
for multivariate data.  For univariate, constrOptim is recommended.  For
multivariate, constrOptim is also recommended as it appears to be faster.}
}
\value{
A named list containing the results of the fit.  beta vector is
equivalent to the mean estimate if x = matrix of 1's, and omega/alpha/nu
are the parameters of the skew-t.  A convergence flag is also returned,
indicating if the solution is a true optimum.
}
\description{
Fits a robust version of the multivariate skew-t, done by bounding the
negative log-likelihood for each observation.  Re-compute the value of k
each time new parameter estimates are available.
}

