% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/runSim.R
\name{runSim}
\alias{runSim}
\title{Run simulation}
\usage{
runSim(n = 100, outPct = 0, outSigma = 0, k = 6:10, fast.k = c(0.99,
  0.98, 0.95, 0.9), xi0 = runif(1, min = -10, max = 10), Omega0 = diag(x =
  10^runif(1, -1, 1), nrow = 1), alpha0 = runif(1, -4, 4), nu0 = 10^runif(1,
  log(4)/log(10), 4), pressure = NULL, type = NULL, restrict = FALSE)
}
\arguments{
\item{n}{Sample size to simulate.}

\item{outPct}{Percent of outliers.  Actual number is round(n*outPct)}

\item{k}{A numeric vector of constants to use in "capping" the likelihood
function.  Each constant is tried in turn, and results are returned for all
values.}

\item{fast.k}{A vector of k values for the fastTLE algorithm.  Each value
should represent the proportion of data to be used in estimating the MLE.}

\item{xi0}{Simulated center parameter of the skew-t to simulate.}

\item{alpha0}{Simulated skewness parameter of the skew-t to simulate.}

\item{nu0}{Simulated heaviness of tails parameter of the skew-t to simulate.}

\item{pressure}{Use Denver station data at this pressure level, and
simulate the skew-t using Ying's code}

\item{type}{The type of skewness to use.  Must be "MVN", "obs", or "EX".}

\item{restrict}{Should outliers be generated in the tail of the skew-t
only?  Only applies if pressure and type are not null.}

\item{outSigma:}{Outliers are created by adding on an error of
N(0,outSigma^2)}

\item{omega0}{Simulated scale parameter of the skew-t to simulate.}
}
\value{
A data.frame containing the results of the simulations.
}
\description{
This function runs a simulation to measure the performance of the robust
and non-robust fitting methods on data where the distribution is known.
}

